test: `nucleus/src/plugins` done
